Fix class level variable issue
The operation configuration values should not be stored in class level
variables. Becasue, the class mediators in the connector operations will
be deployed only once as that is inside the connector operation sequence
template. So, all the same operations that are used multiple times in the
integration flow will share the same class.